결재자 미배정 목록 페이지에서 조사자가 배정된 목록 리스트만 노출되도록 구현. 결재자 배정 기능 구현 필요.

+++ Spring/src/main/java/bitc/full502/final_project_team1/api/web/dto/BuildingListItemDto.java
@@ -1,7 +1,7 @@
 package bitc.full502.final_project_team1.api.web.dto;
 
-import bitc.full502.final_project_team1.core.domain.repository.projection.BuildingListProjection;
 import lombok.*;
+import bitc.full502.final_project_team1.core.domain.repository.projection.BuildingListProjection;
 
 @Getter @Setter @NoArgsConstructor @AllArgsConstructor @Builder
 public class BuildingListItemDto {
@@ -10,7 +10,7 @@ public class BuildingListItemDto {
     private String roadAddress;
 
     private boolean assigned;          // 배정 여부
-    private Integer assignedUserId;
+    private Long assignedUserId;
     private String  assignedUserName;
 
     private String  resultStatus;      // 최신 조사/결재 상태 (e.g. APPROVED/SENT/TEMP/...)
