+++ Spring/src/main/java/bitc/full502/final_project_team1/core/service/SurveyServiceImpl.java
@@ -24,7 +24,7 @@ public class SurveyServiceImpl implements SurveyService {
     private final SurveyResultRepository surveyResultRepository;
 
     @Override
-    public List<AssignedBuildingDto> assigned(Integer userId) {
+    public List<AssignedBuildingDto> assigned(Long userId) {
         List<Object[]> rows = appAssignmentQueryRepository.findAssignedAll(userId);
         return rows.stream()
                 .map(this::toDtoNoDistance)
@@ -32,7 +32,7 @@ public class SurveyServiceImpl implements SurveyService {
     }
 
     @Override
-    public List<AssignedBuildingDto> assignedWithin(Integer userId, double lat, double lng, double radiusKm) {
+    public List<AssignedBuildingDto> assignedWithin(Long userId, double lat, double lng, double radiusKm) {
         double meters = radiusKm * 1000.0;
         List<Object[]> rows = appAssignmentQueryRepository.findAssignedWithin(userId, lat, lng, meters);
         return rows.stream()
@@ -74,7 +74,7 @@ public class SurveyServiceImpl implements SurveyService {
     private String toStr(Object o) { return o == null ? null : o.toString(); }
 
     @Override
-    public AppUserSurveyStatusResponse getStatus(Integer userId) {
+    public AppUserSurveyStatusResponse getStatus(Long userId) {
         Map<String, Long> counts = surveyResultRepository.countGroupByStatus(userId).stream()
                 .collect(Collectors.toMap(SurveyResultRepository.StatusCount::getStatus,
                         SurveyResultRepository.StatusCount::getCnt));
@@ -89,7 +89,7 @@ public class SurveyServiceImpl implements SurveyService {
 
     @Override
     public ListWithStatusResponse<SurveyListItemDto> getListWithStatus(
-            Integer userId, String status, int page, int size
+            Long userId, String status, int page, int size
     ) {
         Page<SurveyResultEntity> p = surveyResultRepository.findByUserAndStatusPage(
                 userId, status, PageRequest.of(page, size));
